pyproject.toml
[project]
name = "gold-intelligence-framework"
version = "0.1.0"
description = "Enterprise Gold Market Analysis Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "dbnomics>=1.0.0",
    "duckdb>=0.9.0",
    "google-cloud-storage>=2.0.0",
    "pyarrow>=12.0.0",
    "dbt-duckdb>=1.7.0",
    "dbt-bigquery>=1.7.0",
    "apache-airflow>=2.7.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
